이메일 인증 완료

+++ src/main/resources/static/js/board.js
@@ -19,38 +19,43 @@ let index = {
 			title:$("#title").val(),
 			content:$("#content").val(),
 		};
-		
-		$.ajax({
-			type: "POST",
-			url: "/api/board", 
-			data: JSON.stringify(data), // http body 데이터
-			contentType: "application/json; charset=utf-8", // body 데이터가 어떤 타입인지 (MIME)
-			dataType: "json" // 요청을 서버로해서 응답이 왔을 때 기본적으로 모든 것이 문자열(생긴게 json이라면) => javascript 오브젝트로 변경
-		}).done(function(resp){ // 성공
-			alert("글 작성이 완료되었습니다.");
-			console.log(resp);
-			location.href="/";
-		}).fail(function(error){ //실패
-			alert(JSON.stringify(error));
-		});  // ajax 통신을 이용해서 3개의 데이터를 json으로 변경하여 insert 요청
-		
+		if(confirm("작성하시겠습니까?")){
+			$.ajax({
+				type: "POST",
+				url: "/api/board", 
+				data: JSON.stringify(data), // http body 데이터
+				contentType: "application/json; charset=utf-8", // body 데이터가 어떤 타입인지 (MIME)
+				dataType: "json" // 요청을 서버로해서 응답이 왔을 때 기본적으로 모든 것이 문자열(생긴게 json이라면) => javascript 오브젝트로 변경
+			}).done(function(resp){ // 성공
+				alert("글 작성이 완료되었습니다.");
+				console.log(resp);
+				location.href="/";
+			}).fail(function(error){ //실패
+				alert(JSON.stringify(error));
+			});  // ajax 통신을 이용해서 3개의 데이터를 json으로 변경하여 insert 요청
+			
+		}
+			
 	},
 	
 	deleteById: function() {
 		let id = $("#id").text();
 		
-		$.ajax({
-			type: "DELETE",
-			url: "/api/board/"+id, 
-			dataType: "json" // 요청을 서버로해서 응답이 왔을 때 기본적으로 모든 것이 문자열(생긴게 json이라면) => javascript 오브젝트로 변경
-		}).done(function(resp){ // 성공
-			alert("삭제가 완료되었습니다.");
-			console.log(resp);
-			location.href="/";
-		}).fail(function(error){ //실패
-			alert(JSON.stringify(error));
-		}); 
+		if(confirm("삭제하시겠습니까?")){
+			$.ajax({
+				type: "DELETE",
+				url: "/api/board/"+id, 
+				dataType: "json" // 요청을 서버로해서 응답이 왔을 때 기본적으로 모든 것이 문자열(생긴게 json이라면) => javascript 오브젝트로 변경
+			}).done(function(resp){ // 성공
+				alert("삭제가 완료되었습니다.");
+				console.log(resp);
+				location.href="/";
+			}).fail(function(error){ //실패
+				alert(JSON.stringify(error));
+			}); 			
+		}
 		
+	
 	},
 	
 	update: function() {
@@ -60,6 +65,8 @@ let index = {
 			title:$("#title").val(),
 			content:$("#content").val(),
 		};
+		
+		if(confirm("수정하시겠습니까?")){
 		$.ajax({
 			type: "PUT",
 			url: "/api/board/" +id, 
@@ -72,8 +79,8 @@ let index = {
 			location.href="/";
 		}).fail(function(error){ //실패
 			alert(JSON.stringify(error));
-		});  
-		
+		});  			
+		}
 	},
 	replysave: function() {
 		let data = {
